html/capability-delegation/delegation-sender-checks.sub.tentative.html

!DOCTYPE html>
<!--
   Tentative due to:
     https://github.com/WICG/capability-delegation
     https://github.com/w3c/webappsec-permissions-policy/issues/475
-->
<title>Capability Delegation sender checks</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/utils.js"></script>

<div>
  Verifies that capability delegation related error checks in <a
  href="https://wicg.github.io/capability-delegation/spec.html#monkey-patch-to-html-initiating-delegation">HTML
  postMessage algorithm</a> are enforced correctly.
</div>

<iframe width="300px" height="50px"
    src="https://{{hosts[alt][www]}}:{{ports[https][0]}}/html/capability-delegation/resources/minimal.html">
</iframe>

<script>
  function getCapabilityDelegationMessageException(frame, origin, capability) {
      try {
          frame.postMessage("any_message", {targetOrigin: origin, delegate: capability});
          return undefined;
      } catch (e) {
          return e;
      }
  }

  const origin = "https://{{hosts[alt][www]}}:{{ports[https][0]}}";

  promise_test(async () => {
      let exception = getCapabilityDelegationMessageException(frames[0], origin, "blah");
      assert_equals(exception.name, "NotSupportedError");
  }, `Delegating an unsupported capability throws an exception`);

  promise_test(async () => {
      let exception = getCapabilityDelegationMessageException(frames[0], "*", "no-op");
      assert_equals(exception.name, "NotAllowedError");
  }, `Delegating to targetOrigin='*' throws an exception`);

  promise_test(async () => {
      let exception = getCapabilityDelegationMessageException(frames[0], origin, "no-op");
      assert_equals(exception.name, "NotAllowedError");
  }, `Delegating without user activation throws an exception`);
</script>
